refactor: extract org management to repository layer
- Create repositories/org_repository.rs
- Subdivide api/orgs.rs into list.rs, crud.rs, members.rs, invitations.rs, settings.rs, logo.rs
- Add UserRepository::get_by_email method for invitation lookups
- Update auth/oauth.rs to use OrgRepository::add_member
- Update api/orgs/invitations.rs to use UserRepository for user queries
- Remove api/orgs.rs and org-related SQL functions from db/queries.rs
