+++ src/openapi.rs
@@ -88,23 +88,23 @@ use utoipa::{Modify, OpenApi};
         crate::api::tags::handle_rename_org_tag,
 
         // Organizations
-        crate::api::orgs::handle_list_user_orgs,
-        crate::api::orgs::handle_create_org,
-        crate::api::orgs::handle_switch_org,
-        crate::api::orgs::handle_get_org,
-        crate::api::orgs::handle_update_org,
-        crate::api::orgs::handle_delete_org,
-        crate::api::orgs::handle_get_org_settings,
-        crate::api::orgs::handle_update_org_settings,
-        crate::api::orgs::handle_remove_member,
-        crate::api::orgs::handle_create_invitation,
-        crate::api::orgs::handle_revoke_invitation,
-        crate::api::orgs::handle_resend_invitation,
-        crate::api::orgs::handle_get_invite_info,
-        crate::api::orgs::handle_accept_invite,
-        crate::api::orgs::handle_upload_org_logo,
-        crate::api::orgs::handle_get_org_logo,
-        crate::api::orgs::handle_delete_org_logo,
+        crate::api::orgs::list::handle_list_user_orgs,
+        crate::api::orgs::crud::handle_create_org,
+        crate::api::orgs::list::handle_switch_org,
+        crate::api::orgs::crud::handle_get_org,
+        crate::api::orgs::crud::handle_update_org,
+        crate::api::orgs::crud::handle_delete_org,
+        crate::api::orgs::settings::handle_get_org_settings,
+        crate::api::orgs::settings::handle_update_org_settings,
+        crate::api::orgs::members::handle_remove_member,
+        crate::api::orgs::invitations::handle_create_invitation,
+        crate::api::orgs::invitations::handle_revoke_invitation,
+        crate::api::orgs::invitations::handle_resend_invitation,
+        crate::api::orgs::invitations::handle_get_invite_info,
+        crate::api::orgs::invitations::handle_accept_invite,
+        crate::api::orgs::logo::handle_upload_org_logo,
+        crate::api::orgs::logo::handle_get_org_logo,
+        crate::api::orgs::logo::handle_delete_org_logo,
 
         // API Keys
         crate::api::keys::handle_create_api_key,
