+++ src/repositories/mod.rs
@@ -18,6 +18,7 @@ pub mod analytics_repository;
 pub mod api_key_repository;
 pub mod billing_repository;
 pub mod blacklist_repository;
+pub mod org_repository;
 pub mod product_repository;
 pub mod report_repository;
 pub mod settings_repository;
@@ -27,6 +28,7 @@ pub use analytics_repository::AnalyticsRepository;
 pub use api_key_repository::ApiKeyRepository;
 pub use billing_repository::BillingRepository;
 pub use blacklist_repository::BlacklistRepository;
+pub use org_repository::OrgRepository;
 pub use product_repository::ProductRepository;
 pub use report_repository::ReportRepository;
 pub use settings_repository::SettingsRepository;
